Moved job methods to more appropriate location.
