+++ session.go
@@ -4,6 +4,7 @@ import (
 	"bufio"
 	"bytes"
 	"fmt"
+	"golang.org/x/crypto/ssh"
 	"io"
 	"sync"
 )
@@ -31,6 +32,17 @@ type Result struct {
 	StdErrStream chan []byte
 	// Written to when a host completes it's work. This does not indicate that all output from StdOutStream or StdErrStream has been read and/or processed.
 	DoneChannel chan struct{}
+}
+
+// getJob determines the type of job and returns the command string
+func getJob(s *ssh.Session, j *Job) string {
+	// Set up remote script
+	if j.Script != nil {
+		s.Stdin = bytes.NewReader(j.Script)
+		return fmt.Sprintf("cat > outfile.sh && chmod +x ./outfile.sh && ./outfile.sh %s && rm ./outfile.sh", j.ScriptArgs)
+	}
+
+	return j.Command
 }
 
 // sshCommand runs an SSH task and returns Result only when the command has finished executing.
